refactor: jobs to tsx, allow reactnode for alt in photo
