+++ pages/jobs/clubs-lead.tsx
@@ -2,16 +2,16 @@ import { BaseStyles, Box, Container, Heading, Text } from 'theme-ui'
 import Head from 'next/head'
 import Nav from '../../components/nav'
 import Meta from '@hackclub/meta'
-import JobDescription from '../../components/jobs/hcb-ops-lead/jd.mdx'
+import JobDescription from '../../components/jobs/clubs-lead/jd.mdx'
 import ForceTheme from '../../components/force-theme'
 
 const Page = () => (
   <>
     <Meta
       as={Head}
-      title="HCB Operations Lead"
-      description="Hack Club is a hiring a HCB Operations Lead as the 7th full-time member of our team in Burlington, Vermont."
-      image="https://workshop-cards.hackclub.com/fiscal-sponsorship%20Ops%20Lead%20%40%20Hack%20Club.png?fontSize=175px&brand=HQ"
+      title="Clubs Lead"
+      description="Hack Club is a hiring a Clubs Lead as the 8th full-time member of our team in Burlington, Vermont."
+      image="https://workshop-cards.hackclub.com/Clubs%20Lead%20%40%20Hack%20Club.png?fontSize=200px&brand=HQ"
     />
     <ForceTheme theme="light" />
     <Nav />
@@ -20,15 +20,15 @@ const Page = () => (
       sx={{
         pt: [5, 6],
         pb: [4, 5],
-        backgroundImage: theme => theme.util.gx('red', 'purple')
+        backgroundImage: (t: any) => t.util.gx('blue', 'purple')
       }}
     >
       <Container sx={{ textAlign: 'center', color: 'white' }}>
         <Heading as="h1" variant="title" mb={30}>
-          HCB Operations Lead
+          Clubs Lead @ Hack Club
         </Heading>
         <Text variant="headline" sx={{ fontWeight: 400 }}>
-          Closed as of March 5th, 2021.
+          New job open as of May 18th, 2021.
         </Text>
       </Container>
     </Box>
