+++ pages/jobs/customer-success-lead.tsx
@@ -2,16 +2,16 @@ import { BaseStyles, Box, Container, Heading, Text } from 'theme-ui'
 import Head from 'next/head'
 import Nav from '../../components/nav'
 import Meta from '@hackclub/meta'
-import JobDescription from '../../components/jobs/executive-producer/jd.mdx'
+import JobDescription from '../../components/jobs/customer-success-lead/jd.mdx'
 import ForceTheme from '../../components/force-theme'
 
 const Page = () => (
   <>
     <Meta
       as={Head}
-      title="Executive Producer"
-      description="Hack Club is actively looking to hire its first Executive Producer to tell its story across various platforms, including social media, newsletters, op-docs, blog/Op-Ed posts, direct emails, and curated partnerships with major tech companies."
-      image="https://workshop-cards.hackclub.com/Executive%20Producer%20%40%20Hack%20Club.png?theme=light&md=1&fontSize=175px&caption="
+      title="Customer Success Lead"
+      description="Hack Club is a hiring a Customer Success Lead to join our team in Shelburne, Vermont."
+      image="https://workshop-cards.hackclub.com/Customer%20Success%20Lead%20%40%20Hack%20Club.png?fontSize=175px&brand=HQ"
     />
     <ForceTheme theme="light" />
     <Nav />
@@ -20,15 +20,15 @@ const Page = () => (
       sx={{
         pt: [5, 6],
         pb: [4, 5],
-        backgroundImage: theme => theme.util.gx('green', 'blue')
+        backgroundImage: (t: any) => t.util.gx('red', 'yellow')
       }}
     >
       <Container sx={{ textAlign: 'center', color: 'white' }}>
         <Heading as="h1" variant="title" mb={30}>
-          Executive Producer
+          Customer Success Lead @&nbsp;Hack&nbsp;Club
         </Heading>
         <Text variant="headline" sx={{ fontWeight: 400 }}>
-          New job open as of October 2022.
+          This position has been filled as of June 14th, 2022.
         </Text>
       </Container>
     </Box>
