+++ pages/jobs/education-engineer.tsx
@@ -2,16 +2,16 @@ import { BaseStyles, Box, Container, Heading, Text } from 'theme-ui'
 import Head from 'next/head'
 import Nav from '../../components/nav'
 import Meta from '@hackclub/meta'
-import JobDescription from '../../components/jobs/customer-success-lead/jd.mdx'
+import JobDescription from '../../components/jobs/education-engineer/jd.mdx'
 import ForceTheme from '../../components/force-theme'
 
 const Page = () => (
   <>
     <Meta
       as={Head}
-      title="Customer Success Lead"
-      description="Hack Club is a hiring a Customer Success Lead to join our team in Shelburne, Vermont."
-      image="https://workshop-cards.hackclub.com/Customer%20Success%20Lead%20%40%20Hack%20Club.png?fontSize=175px&brand=HQ"
+      title="Education Engineer"
+      description="Hack Club is a hiring for an engineer/educator to create and lead on technical projects for Hack Clubbers, with a strong focus on girls at Hack Club."
+      image="https://workshop-cards.hackclub.com/Education%20Engineer%20%40%20Hack%20Club.png?fontSize=180px&brand=HQ"
     />
     <ForceTheme theme="light" />
     <Nav />
@@ -20,16 +20,14 @@ const Page = () => (
       sx={{
         pt: [5, 6],
         pb: [4, 5],
-        backgroundImage: theme => theme.util.gx('red', 'yellow')
+        backgroundImage: (t: any) => t.util.gx('red', 'orange')
       }}
     >
       <Container sx={{ textAlign: 'center', color: 'white' }}>
-        <Heading as="h1" variant="title" mb={30}>
-          Customer Success Lead @&nbsp;Hack&nbsp;Club
+        <Heading as="h1" variant="title">
+          Education Engineer @&nbsp;Hack&nbsp;Club
         </Heading>
-        <Text variant="headline" sx={{ fontWeight: 400 }}>
-          This position has been filled as of June 14th, 2022.
-        </Text>
+        <Text variant="headline">New job open as of November 20th, 2021.</Text>
       </Container>
     </Box>
     <Container
