+++ pages/jobs/executive-assistant.tsx
@@ -2,16 +2,16 @@ import { BaseStyles, Box, Container, Heading, Text } from 'theme-ui'
 import Head from 'next/head'
 import Nav from '../../components/nav'
 import Meta from '@hackclub/meta'
-import JobDescription from '../../components/jobs/clubs-lead/jd.mdx'
+import JobDescription from '../../components/jobs/executive-assistant/jd.mdx'
 import ForceTheme from '../../components/force-theme'
 
 const Page = () => (
   <>
     <Meta
       as={Head}
-      title="Clubs Lead"
-      description="Hack Club is a hiring a Clubs Lead as the 8th full-time member of our team in Burlington, Vermont."
-      image="https://workshop-cards.hackclub.com/Clubs%20Lead%20%40%20Hack%20Club.png?fontSize=200px&brand=HQ"
+      title="Executive Assistant"
+      description="Hack Club is a hiring an Executive Assistant to join our team in Shelburne, Vermont."
+      image="https://workshop-cards.hackclub.com/Executive%20Assistant%20%40%20Hack%20Club.png?fontSize=180px&brand=HQ"
     />
     <ForceTheme theme="light" />
     <Nav />
@@ -20,15 +20,15 @@ const Page = () => (
       sx={{
         pt: [5, 6],
         pb: [4, 5],
-        backgroundImage: theme => theme.util.gx('blue', 'purple')
+        backgroundImage: (t: any) => t.util.gx('blue', 'green')
       }}
     >
       <Container sx={{ textAlign: 'center', color: 'white' }}>
         <Heading as="h1" variant="title" mb={30}>
-          Clubs Lead @ Hack Club
+          Executive Assistant @&nbsp;Hack&nbsp;Club
         </Heading>
         <Text variant="headline" sx={{ fontWeight: 400 }}>
-          New job open as of May 18th, 2021.
+          New job open as of January 24th, 2022.
         </Text>
       </Container>
     </Box>
