+++ pages/jobs/hcb-ops-associate.tsx
@@ -2,16 +2,16 @@ import { BaseStyles, Box, Container, Heading, Text } from 'theme-ui'
 import Head from 'next/head'
 import Nav from '../../components/nav'
 import Meta from '@hackclub/meta'
-import JobDescription from '../../components/jobs/brand-director/jd.mdx'
+import JobDescription from '../../components/jobs/hcb-ops-associate/jd.mdx'
 import ForceTheme from '../../components/force-theme'
 
 const Page = () => (
   <>
     <Meta
       as={Head}
-      title="Brand Director"
-      description="Hack Club is a hiring a Brand Director as the 7th full-time member of our team in Burlington, Vermont."
-      image="https://workshop-cards.hackclub.com/Brand%20Director%20%40%20Hack%20Club.png?fontSize=175px&brand=HQ"
+      title="HCB Operations Associate"
+      description="Hack Club is a hiring a HCB Operations Associate as the 8th full-time member of our team in Burlington, Vermont."
+      image="https://workshop-cards.hackclub.com/fiscal-sponsorship%20Ops%20Associate%20%40%20Hack%20Club.png?fontSize=175px&brand=HQ"
     />
     <ForceTheme theme="light" />
     <Nav />
@@ -20,15 +20,15 @@ const Page = () => (
       sx={{
         pt: [5, 6],
         pb: [4, 5],
-        backgroundImage: theme => theme.util.gx('blue', 'purple')
+        backgroundImage: (t: any) => t.util.gx('yellow', 'red')
       }}
     >
       <Container sx={{ textAlign: 'center', color: 'white' }}>
         <Heading as="h1" variant="title" mb={30}>
-          Hack Club Brand Director
+          HCB Operations Associate
         </Heading>
         <Text variant="headline" sx={{ fontWeight: 400 }}>
-          Tell the story of Hack Club. New job open as of February 8th, 2021.
+          New job open as of May 21st, 2021.
         </Text>
       </Container>
     </Box>
