+++ pages/jobs/hcb-ops-lead.tsx
@@ -2,16 +2,16 @@ import { BaseStyles, Box, Container, Heading, Text } from 'theme-ui'
 import Head from 'next/head'
 import Nav from '../../components/nav'
 import Meta from '@hackclub/meta'
-import JobDescription from '../../components/jobs/education-engineer/jd.mdx'
+import JobDescription from '../../components/jobs/hcb-ops-lead/jd.mdx'
 import ForceTheme from '../../components/force-theme'
 
 const Page = () => (
   <>
     <Meta
       as={Head}
-      title="Education Engineer"
-      description="Hack Club is a hiring for an engineer/educator to create and lead on technical projects for Hack Clubbers, with a strong focus on girls at Hack Club."
-      image="https://workshop-cards.hackclub.com/Education%20Engineer%20%40%20Hack%20Club.png?fontSize=180px&brand=HQ"
+      title="HCB Operations Lead"
+      description="Hack Club is a hiring a HCB Operations Lead as the 7th full-time member of our team in Burlington, Vermont."
+      image="https://workshop-cards.hackclub.com/fiscal-sponsorship%20Ops%20Lead%20%40%20Hack%20Club.png?fontSize=175px&brand=HQ"
     />
     <ForceTheme theme="light" />
     <Nav />
@@ -20,14 +20,16 @@ const Page = () => (
       sx={{
         pt: [5, 6],
         pb: [4, 5],
-        backgroundImage: theme => theme.util.gx('red', 'orange')
+        backgroundImage: (t: any) => t.util.gx('red', 'purple')
       }}
     >
       <Container sx={{ textAlign: 'center', color: 'white' }}>
-        <Heading as="h1" variant="title">
-          Education Engineer @&nbsp;Hack&nbsp;Club
+        <Heading as="h1" variant="title" mb={30}>
+          HCB Operations Lead
         </Heading>
-        <Text variant="headline">New job open as of November 20th, 2021.</Text>
+        <Text variant="headline" sx={{ fontWeight: 400 }}>
+          Closed as of March 5th, 2021.
+        </Text>
       </Container>
     </Box>
     <Container
