+++ pages/jobs/lead-hacker.tsx
@@ -22,7 +22,7 @@ const Page = () => (
       sx={{
         pt: [5, 6],
         pb: [4, 5],
-        backgroundImage: theme => theme.util.gx('purple', 'red')
+        backgroundImage: (t: any) => t.util.gx('purple', 'red')
       }}
     >
       <Container sx={{ textAlign: 'center', color: 'white' }}>
