+++ pages/jobs/philanthropy-communications-lead.tsx
@@ -2,16 +2,16 @@ import { BaseStyles, Box, Container, Heading, Text } from 'theme-ui'
 import Head from 'next/head'
 import Nav from '../../components/nav'
 import Meta from '@hackclub/meta'
-import JobDescription from '../../components/jobs/vp-donor-engagement/jd.mdx'
+import JobDescription from '../../components/jobs/philanthropy-communications-lead/jd.mdx'
 import ForceTheme from '../../components/force-theme'
 
 const Page = () => (
   <>
     <Meta
       as={Head}
-      title="VP of Donor Engagement"
-      description="Hack Club is a hiring a VP of Donor Engagement to join our team in Shelburne, Vermont."
-      image="https://workshop-cards.hackclub.com/VP%2C%20Donor%20Engagment%20%40%20Hack%20Club.png?fontSize=160px&brand=HQ"
+      title="Philanthropy Communications Lead"
+      description="Hack Club is a hiring a Philanthropy Communications Lead to join our team in Shelburne, Vermont."
+      image="https://workshop-cards.hackclub.com/Philanthropy%20Communications%20Lead%20%40%20Hack%20Club.png?fontSize=200px&brand=HQ"
     />
     <ForceTheme theme="light" />
     <Nav />
@@ -20,15 +20,15 @@ const Page = () => (
       sx={{
         pt: [5, 6],
         pb: [4, 5],
-        backgroundImage: theme => theme.util.gx('pink', 'yellow')
+        backgroundImage: (t: any) => t.util.gx('green', 'blue')
       }}
     >
       <Container sx={{ textAlign: 'center', color: 'white' }}>
         <Heading as="h1" variant="title" mb={30}>
-          Vice President, Donor Engagement @&nbsp;Hack&nbsp;Club
+          Philanthropy Communications Lead @&nbsp;Hack&nbsp;Club
         </Heading>
         <Text variant="headline" sx={{ fontWeight: 400 }}>
-          New job open as of Feburary 16th, 2022.
+          This position has been filled as of May 3rd, 2022.
         </Text>
       </Container>
     </Box>
